refactor(Level): New GameMap class, map transition
screens/level.py
- Player is now created prior to map setup
- a new transition similar to the day transition has been added and will be played when the Player is traversing between maps
- all methods related to setting up an individual map have been extracted to the new GameMap class in screens/game_map.py
- the Player will now be positioned depending on which map they came from, and will fall back to a default spawn point

screens/game_map.py
- now contains all logic regarding map setup
- map layers are no longer loaded by their layer name, but also depending on whether it is an object or tile layer, and which layer property has been set. This ensures that decorative layers that have been split into multiple layers for better readability / maintainability are now correctly loaded across all maps

sprites/base.py
- CollideableMapObject and AnimatedSprite now have None as default for the groups attribute

enums.py
- merged Layer.LOWER_GROUND and UPPER_GROUND into Layer.GROUND

sprites/entities/player.py
- removed font attribute from Player
- added Player.teleport method which moves the Player rect without checking for collision

npc/setup.py
- renamed AIData.setup to AIData.update, which can now be called multiple times. Behaviour trees and pathfinding finders will only be initialised once

overlay/soil.py
- soil tiles should now be created by calling SoilLayer.create_soil_tiles after initialising a new SoilLayer instance

- fixed flake8 and type hinting issues across multiple files

+++ src/npc/npc.py
@@ -25,7 +25,7 @@ class NPC(NPCBase):
             groups: tuple[pygame.sprite.Group, ...],
             collision_sprites: pygame.sprite.Group,
             apply_tool: Callable[
-                [FarmingTool, tuple[int, int], Character], None
+                [FarmingTool, tuple[float, float], Character], None
             ],
             soil_layer: SoilLayer,
             emote_manager: NPCEmoteManager
